Agenda view showing unclocked old gcal.org entries
This view will make it easy to delete all these old entries, which do
nothing but take up space and make Org slower. To make this easier, I've
replace the `org-agenda-kill` command with an override to move to the
next line if the current line doesn't contain an Org headline.