Add option to start org-pomodoro breaks without locking screen
By default I lock the screen when the Pomodoro finishes, but I want to
be able to suppress that when I want to take a short break while still
working at the computer (if I want to force myself to start a Pomodoro
in 5 minutes, for example). I've added a function to start a short
break as well.