Add option to schedule break reminder alarm
This is an alarm to allow me to remember to do stuff around the house
while I'm on a pomodoro break. As part of this, refactor the alarm
handling code and fix several bugs, including one where the
alarm event IDs would not be set to nil after removing the event,
because I was using `async-start-process`, which executes everything
in a separate process.