fix: org-super-agenda: recurring events fix.
In current versions of org-super-agenda, since I exclude the parent
recurring event from the agenda, a child recurring event of the same
name and time will also be excluded. For now, then, disable excluding
the parent recurring event. In that case, both events will appear, but
that’s the best that can be done right now.