my-org-git-sync: allow push
This ensures that the Org git repo continues to be updated, which is
something that hasn't been happening for a while (since I've needed to
manually call `org-syncup-full` without disabling push). Now that I've
switched to using Syncthing for the Org repos, running a noninteractive
push should be much more failsafe than before.